Modulo3/atividades/11-10-2024/requirements_uninstallFlet.txt
annotated-types==0.7.0
anyio==4.6.0
arrow==1.3.0
bidict==0.23.1
binaryornot==0.4.4
blinker==1.8.2
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cookiecutter==2.6.0
exceptiongroup==1.2.2
fastapi==0.115.0
Flask==3.0.3
Flask-SocketIO==5.4.1
flet-core==0.24.1
flet-runtime==0.24.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
oauthlib==3.2.2
packaging==23.2
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-slugify==8.0.4
python-socketio==5.11.4
PyYAML==6.0.2
qrcode==7.4.2
repath==0.9.0
requests==2.32.3
rich==13.9.2
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
text-unidecode==1.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.1
uvloop==0.20.0
watchdog==4.0.2
watchfiles==0.24.0
websockets==13.1
Werkzeug==3.0.4
wsproto==1.2.0
